requirements.txt
pandas
openai
tiktoken
bytesbufio
requests
pillow
-e .